Rename sendmail to notify
Make the module name less confusing when adding more notification types.
